pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "towervision"
version = "0.1.0"
description = "Scaffold para benchmark de detecção de isoladores e anomaly detection no projeto Tower Vision."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
benchmark = [
  "accelerate>=1.13.0,<2",
  "anomalib>=2.2.0,<3",
  "matplotlib>=3.8.0,<4",
  "numpy>=2.2.0,<3",
  "pandas>=2.2.0,<3",
  "pycocotools>=2.0.11,<3",
  "rich>=14.0.0,<15",
  "torch>=2.7.0,<2.8",
  "torchvision>=0.22.0,<0.23",
  "transformers>=4.57.0,<5",
  "ultralytics>=8.4.0,<9",
]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/towervision"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
